LavaDelta 3.0.0
Name change from LavaSolution to LavaDelta. Changed the entire UI of the project, responsive and desktop version. Also added legends to the carousel images, and added more images, changed the way the dishwashers looked in photoshop. In addition ive installed react-scroll which made smoother scrolls in the main page, but now has began causing troubles with the sites navigation